# websocket: optimize masking using WebAssembly