--- a/lib/web/websocket/connection.js
+++ b/lib/web/websocket/connection.js
@@ -26,6 +26,19 @@ try {
 
 }
 
+/** @type {import('ws-masking')} */
+let wsm
+
+try {
+  wsm = require('ws-masking')
+} catch {
+  // do noting
+}
+
+let wsmInitialized = false
+
+let wsmExperimentalWarned = false
+
 /**
  * @see https://websockets.spec.whatwg.org/#concept-websocket-establish
  * @param {URL} url
@@ -104,7 +117,20 @@ function establishWebSocketConnection (url, protocols, client, ws, onEstablish, 
     request,
     useParallelQueue: true,
     dispatcher: options.dispatcher,
-    processResponse (response) {
+    async processResponse (response) {
+      if (wsm !== undefined && process.env.UNDICI_USE_WS_MASKING === 'true') {
+        // init ws-masking wasm.
+        if (!wsmExperimentalWarned) {
+          wsmExperimentalWarned = true
+          process.emitWarning('WebSocket optimization using `ws-masking` is experimental, expect them to change at any time', {
+            code: 'UNDICI-WS-MASKING'
+          })
+        }
+        if (!wsmInitialized) {
+          await wsm.initialize()
+          wsmInitialized = true
+        }
+      }
       // 1. If response is a network error or its status is not 101,
       //    fail the WebSocket connection.
       if (response.type === 'error' || response.status !== 101) {