--- a/lib/web/websocket/frame.js
+++ b/lib/web/websocket/frame.js
@@ -2,6 +2,15 @@
 
 const { maxUnsigned16Bit } = require('./constants')
 
+/** @type {import('ws-masking')} */
+let wsm
+
+try {
+  wsm = require('ws-masking')
+} catch {
+  // do noting
+}
+
 const BUFFER_SIZE = 16386
 
 /** @type {import('crypto')} */
@@ -29,7 +38,7 @@ function generateMask () {
     bufIdx = 0
     crypto.randomFillSync((buffer ??= Buffer.allocUnsafe(BUFFER_SIZE)), 0, BUFFER_SIZE)
   }
-  return [buffer[bufIdx++], buffer[bufIdx++], buffer[bufIdx++], buffer[bufIdx++]]
+  return buffer.slice(bufIdx, (bufIdx += 4))
 }
 
 class WebsocketFrameSend {
@@ -83,8 +92,19 @@ class WebsocketFrameSend {
     buffer[1] |= 0x80 // MASK
 
     // mask body
-    for (let i = 0; i < bodyLength; ++i) {
-      buffer[offset + i] = frameData[i] ^ maskKey[i & 3]
+    if (wsm !== undefined && process.env.UNDICI_USE_WS_MASKING === 'true') {
+      if (bodyLength >= 64) {
+        // See https://github.com/tsctx/ws-masking
+        // use wasm implementation.
+        wsm.mask(frameData, maskKey, buffer, offset, bodyLength)
+      } else {
+        // use fast pure-js implementation for small buffer.
+        wsm.js.mask(frameData, maskKey, buffer, offset, bodyLength)
+      }
+    } else {
+      for (let i = 0; i < bodyLength; ++i) {
+        buffer[offset + i] = frameData[i] ^ maskKey[i & 3]
+      }
     }
 
     return buffer