Closes #3265 

I have created `ws-masking` which allows fast masking in WebAssembly.

I marked this feature as experimental and not available without a flag.

Package details are in https://github.com/tsctx/ws-masking

Script: https://github.com/nodejs/undici/blob/e0e786fabdd5063fb3986495a4bcdb82b5af5e0d/benchmarks/websocket/websocket-send-buffer.mjs

- main

```
• send
------------------------------------------------- -----------------------------
ws          1'958 µs/iter   (1'672 µs … 6'897 µs)  1'956 µs  4'460 µs  6'897 µs
undici      1'926 µs/iter   (1'667 µs … 6'695 µs)  1'942 µs  3'954 µs  6'695 µs

summary for send
  undici
   1.02x faster than ws
```

- this PR (under flag)

```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
• send
------------------------------------------------- -----------------------------
ws          1'880 µs/iter   (1'587 µs … 6'713 µs)  1'870 µs  4'768 µs  6'713 µs
undici        901 µs/iter     (639 µs … 5'268 µs)    917 µs  3'044 µs  5'268 µs

summary for send
  undici
   2.09x faster than ws
```