@KhafraDev 2024-05-17
@lpinca I'm not sure if you'd be interested in this or not.