@lpinca 2024-05-18
It is nice for portability but `bufferutil` seems to be faster in all cases.