@tsctx 2024-05-18
Since it must be copied to the memory area, bufferutil, which does not copy, has the advantage.