@lpinca 2024-05-18
Now that `WebSocket` is in Node.js core, we could add a `Buffer.mask()` and `Buffer.unMask()` implementation using SIMD instructions and exposing it for public usage similarly to how it was done for `Buffer.isUtf8()`.

cc: @lemire @anonrig