@tsctx 2024-05-19
There is a big wall...
* Node
```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
• send
------------------------------------------------- -----------------------------
undici      1'873 µs/iter   (1'530 µs … 4'230 µs)  1'946 µs  3'709 µs  4'230 µs

summary for send
  undici
```
* Node (with PR)
```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
• send
------------------------------------------------- -----------------------------
undici        805 µs/iter     (500 µs … 3'027 µs)    851 µs  1'434 µs  3'027 µs

summary for send
  undici
```
* Node (with zero mask)
```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
• send
------------------------------------------------- -----------------------------
undici        726 µs/iter     (454 µs … 2'701 µs)    819 µs  1'769 µs  2'701 µs

summary for send
  undici
```
* Bun
```
benchmark      time (avg)             (min … max)       p75       p99      p999     
------------------------------------------------- -----------------------------     
• send
------------------------------------------------- -----------------------------     
global        380 µs/iter     (302 µs … 1'812 µs)    371 µs  1'426 µs  1'801 µs

summary for send
  global
```
* Deno
 ```
benchmark      time (avg)             (min … max)       p75       p99      p999     
------------------------------------------------- -----------------------------     
• send
------------------------------------------------- -----------------------------     
global        480 µs/iter     (388 µs … 1'120 µs)    499 µs    738 µs    986 µs

summary for send
  global
```