@lemire 2024-05-19
@lpinca Can you provide the specification for the functions?