@KhafraDev 2024-05-19
https://datatracker.ietf.org/doc/html/rfc6455#section-5.3