@lpinca 2024-05-19
@lemire for a possible API see https://github.com/websockets/bufferutil?tab=readme-ov-file#bufferutilmasksource-mask-output-offset-length. The last 3 parameters can be optional.